pyproject.toml
[project]
name = "smart-slides-automl"
version = "0.1.0"
description = "AutoML service for Generative AI - Auto-tuning LLM pipelines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-pptx>=1.0.2",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "google-generativeai>=0.3.0",
    "aiohttp>=3.9.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=8.2.0",
    "loguru>=0.7.0",
    "numpy>=1.24.0",
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
